fix: refuse NSPI wire values too wide for the field they land in
The restriction discriminant and the relational operator arrive in 32-bit
fields and are stored in 8 bits, and the PROPERTY_VALUE type check
compared only the low half of a 32-bit type field. Each let a malformed
request through as a well-formed one: type 0x100 read as AND, relop 0x100
read as the first comparison, and 0x0001001F passed the check for
PtUnicode. All three now refuse the value instead of truncating it.

+++ internal/nspi/rpccodec.go
@@ -310,7 +310,9 @@ func pullPropValNDR(p *ndr.Pull) (mapi.TaggedPropVal, error) {
 		return tv, err
 	}
 	tv.Tag = mapi.PropTag(tag)
-	if uint16(ptype) != uint16(mapi.PropTag(tag).Type()) {
+	// Compare the full 32 bits the wire carried. PropType is 16 bits, so a
+	// truncating comparison ignored whatever the client put in the high half.
+	if ptype != uint32(mapi.PropTag(tag).Type()) {
 		return tv, fmt.Errorf("%w: property value type %#x != tag type %#x", ndr.ErrFormat, ptype, uint32(mapi.PropTag(tag).Type()))
 	}
 	switch mapi.PropTag(tag).Type() {
